+++ packages/tokens/CHANGELOG.md
@@ -1,4 +1,10 @@
-# @vponte-ignite/tokens
+# @ignite-vponte/tokens
+
+## 4.0.0
+
+### Major Changes
+
+- change package
 
 ## 3.0.0