+++ packages/ts-config/CHANGELOG.md
@@ -1,4 +1,10 @@
-# @vponte-ignite/ts-config
+# @ignite-vponte/ts-config
+
+## 3.0.0
+
+### Major Changes
+
+- change package
 
 ## 2.0.0